backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
openai>=1.59.0
pyyaml>=6.0.2
python-dotenv>=1.0.1
qrcode[pil]>=8.0
